Move ER-specific artifact tests
